Module EST LIBRE OEEE
Module peut se fixer sur toutes les entités qui ont un moduleManager
Une Ent peut pas prendre de collisions de son propre module
Visuel pour liaison de module
Clean du code et lissage  des dysfonctionnements

+++ RonTech/Z_GameManager.pde
@@ -91,7 +91,7 @@ public class GameManager {
 
 class Time {
 
-  double deltaFrames;
+  float deltaFrames;
 
   float getDeltaFrames() {
     deltaFrames = 1 / frameRate;
